+++ main.go
@@ -1,103 +1,34 @@
 package main
 
 import (
-    "encoding/json"
-    "fmt"
-    "log"
-    "net/http"
-    "os"
+	"fmt"
+	"html"
+	"log"
+	"net/http"
+	"os"
 
-    "github.com/gorilla/mux"
+	"github.com/gorilla/mux"
+	"github.com/weisswurstSystems/WWM-BB/security"
 )
 
-// Demo code.....
-type Person struct {
-    ID        string   `json:"id,omitempty"`
-    Firstname string   `json:"firstname,omitempty"`
-    Lastname  string   `json:"lastname,omitempty"`
-    Address   *Address `json:"address,omitempty"`
-}
-
-type Address struct {
-    City  string `json:"city,omitempty"`
-    State string `json:"state,omitempty"`
-}
-
-var people []Person
-
-func GetPersonEndpoint(w http.ResponseWriter, req *http.Request) {
-    params := mux.Vars(req)
-    for _, item := range people {
-        if item.ID == params["id"] {
-            json.NewEncoder(w).Encode(item)
-            return
-        }
-    }
-    json.NewEncoder(w).Encode(&Person{})
-}
-
-func GetPeopleEndpoint(w http.ResponseWriter, req *http.Request) {
-    json.NewEncoder(w).Encode(people)
-}
-
-func CreatePersonEndpoint(w http.ResponseWriter, req *http.Request) {
-    params := mux.Vars(req)
-    var person Person
-    _ = json.NewDecoder(req.Body).Decode(&person)
-    person.ID = params["id"]
-    people = append(people, person)
-    json.NewEncoder(w).Encode(people)
-}
-
-func DeletePersonEndpoint(w http.ResponseWriter, req *http.Request) {
-    params := mux.Vars(req)
-    for index, item := range people {
-        if item.ID == params["id"] {
-            people = append(people[:index], people[index+1:]...)
-            break
-        }
-    }
-    json.NewEncoder(w).Encode(people)
-}
-
-//... bis hierhin
-
-func basicAuthHandler(realm string, next http.HandlerFunc) http.HandlerFunc {
-    return func(w http.ResponseWriter, r *http.Request) {
-        if checkBasicAuth(r) {
-            next(w, r)
-            return
-        }
-
-        w.Header().Set("WWW-Authenticate", fmt.Sprintf(`Basic realm="%s"`, realm))
-        w.WriteHeader(401)
-        w.Write([]byte("401 Unauthorized\n"))
-    }
-}
-
-func checkBasicAuth(r *http.Request) bool {
-    u, p, ok := r.BasicAuth()
-    if !ok {
-        return false
-    }
-    //TODO Datenbankabfrage
-    return u == "user" && p == "user"
+func GetHelloWorldEndpoint(w http.ResponseWriter, req *http.Request) {
+	fmt.Fprintf(w, "Hello, %q", html.EscapeString(req.URL.Path))
 }
 
 func main() {
-    router := mux.NewRouter()
-    people = append(people, Person{ID: "1", Firstname: "Nic", Lastname: "Raboy", Address: &Address{City: "Dublin", State: "CA"}})
-    people = append(people, Person{ID: "2", Firstname: "Maria", Lastname: "Raboy"})
-    router.HandleFunc("/people", GetPeopleEndpoint).Methods("GET")
-    router.HandleFunc("/people/{id}", basicAuthHandler("Please login to see the details of a person", GetPersonEndpoint)).Methods("GET")
-    router.HandleFunc("/people/{id}", basicAuthHandler("Please login to create a person", CreatePersonEndpoint)).Methods("POST")
-    router.HandleFunc("/people/{id}", basicAuthHandler("Please login to delete a person", DeletePersonEndpoint)).Methods("DELETE")
-    port := os.Getenv("PORT")
-    if len(port) == 0 {
-        log.Printf("Starting on port 8080")
-        log.Fatal(http.ListenAndServe(":8080", router))
-    } else {
-        log.Printf("Starting on port %v", port)
-        log.Fatal(http.ListenAndServe(":"+port, router))
-    }
+	router := mux.NewRouter()
+	// unsecured endpoints
+	router.HandleFunc("/", GetHelloWorldEndpoint).Methods("GET")
+
+	// secured endpoints
+	router.HandleFunc("/secured", security.DefaultAuthenticationHandler("Please login to see the details of a person", GetHelloWorldEndpoint)).Methods("GET")
+
+	port := os.Getenv("PORT")
+	if len(port) == 0 {
+		log.Printf("Starting on port 8080")
+		log.Fatal(http.ListenAndServe(":8080", router))
+	} else {
+		log.Printf("Starting on port %v", port)
+		log.Fatal(http.ListenAndServe(":"+port, router))
+	}
 }
